Make same path copy/move a no-op

+++ src/ObsAdapter.php
@@ -166,6 +166,10 @@ class ObsAdapter implements FilesystemAdapter, PublicUrlGenerator, ChecksumProvi
 
     public function move(string $source, string $destination, Config $config): void
     {
+        if ($source === $destination) {
+            return;
+        }
+
         try {
             $this->copy($source, $destination, $config);
             $this->delete($source);
@@ -176,6 +180,10 @@ class ObsAdapter implements FilesystemAdapter, PublicUrlGenerator, ChecksumProvi
 
     public function copy(string $source, string $destination, Config $config): void
     {
+        if ($source === $destination) {
+            return;
+        }
+
         try {
             /** @var string|null $visibility */
             $visibility = $config->get(Config::OPTION_VISIBILITY);
